refactor: extract IRODS Downloader from IRODS
